refactor: move DOM utils to component-base
