+++ packages/component-base/test/dom-utils.test.js
@@ -1,6 +1,6 @@
 import { expect } from '@esm-bundle/chai';
 import { defineCE, fixtureSync } from '@vaadin/testing-helpers';
-import { getAncestorRootNodes } from '../src/dom-utils.js';
+import { addValueToAttribute, getAncestorRootNodes, removeValueFromAttribute } from '../src/dom-utils.js';
 
 describe('dom-utils', () => {
   describe('getAncestorRootNodes', () => {
@@ -34,6 +34,45 @@ describe('dom-utils', () => {
       expect(nodes).to.have.lengthOf(2);
       expect(nodes[0]).to.equal(element.shadowRoot);
       expect(nodes[1]).to.equal(document);
+    });
+  });
+
+  describe('addValueToAttribute', () => {
+    let element;
+
+    beforeEach(() => {
+      element = document.createElement('div');
+    });
+
+    it('should add a value to an attribute', () => {
+      addValueToAttribute(element, 'aria-labelledby', 'label-id');
+      expect(element.getAttribute('aria-labelledby')).to.equal('label-id');
+
+      addValueToAttribute(element, 'aria-labelledby', 'error-id');
+      expect(element.getAttribute('aria-labelledby')).to.equal('label-id error-id');
+    });
+
+    it('should not duplicate values in the attribute', () => {
+      addValueToAttribute(element, 'aria-labelledby', 'label-id');
+      addValueToAttribute(element, 'aria-labelledby', 'label-id');
+      expect(element.getAttribute('aria-labelledby')).to.equal('label-id');
+    });
+  });
+
+  describe('removeValueFromAttribute', () => {
+    let element;
+
+    beforeEach(() => {
+      element = document.createElement('div');
+      element.setAttribute('aria-labelledby', 'label-id error-id');
+    });
+
+    it('should remove a value from an attribute', () => {
+      removeValueFromAttribute(element, 'aria-labelledby', 'error-id');
+      expect(element.getAttribute('aria-labelledby')).to.equal('label-id');
+
+      removeValueFromAttribute(element, 'aria-labelledby', 'label-id');
+      expect(element.hasAttribute('aria-labelledby')).to.be.false;
     });
   });
 });
